+++ CHANGELOG.md
@@ -29,6 +29,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * Increased tvs_maxNumberOfSensors and added check on the number of sensors in stats_tovs file (#404 and !390)
  * Fix the publishing of the program `midas.monitor.Abs` in SSM domain (#410 and !388)
 
+### Removed
+
+ * Removed namelist variables from `NAMBGCK` (#411 and !391).  The variables removed are:
+   * `maxNumSat`
+   * `maxNumTest`
+   * `maxNumChan`
+   * `channelOffset`
+
 ## [3.6.0-a1]
 
 ### Added